Assert Character set equal

Een functie om checkmate::assert_set_equal te gebruiken door een dataframe op te geven, de kolomnamen in een vector en de kolomwaarden waarmee gecheckt wordt.

assert_character_set_equal(
  Kolom,
  df,
  Kolomwaarden,
  prefix_kolom = NULL,
  add,
  ...
)

Arguments

Kolom

een Character vector of string met de kolomnaam die getest moet worden

df

het dataframe dat de kolom bevat

Kolomwaarden

de mogelijke waarden die de geteste kolom kan bevatten

prefix_kolom

standaard NULL. Als deze opgegeven wordt, wordt in de assertion message een tekst voor de variabelenaaam geplakt.

add

Assertion collectie om de assertion-berichten aan toe te voegen

...

De overige parameters worden meegegeven aan de functie assert_set_equal

See also