Assert naming

Functie voor assertions voor een dataframe.

Usage

assert_naming(
  df,
  Naming,
  Naam_bestand,
  Maak_nieuwe_kolomwaarden = F,
  Assert_met_veldnaam = F,
  Bericht = NULL,
  fail = "stop"
)

Arguments

df

Dataframe waar over geassert wordt

Naming

Het bijbehorende documentatie bestand met assert informatie voor het dataframe

Naam_bestand

De naam van het databestand wat bekeken wordt. Deze naam moet uniek zijn in het hele project

Maak_nieuwe_kolomwaarden

True als de functie nog een andere functie, build_kolomwaarden, moet aanroepen om een kolommen bestand te maken waarmee subset/list assertions gedaan kunnen worden.

Assert_met_veldnaam

True als de assertions gedaan worden op de vertaalde namen, dus Veldnaam ipv Veldnaam_export

Bericht

Bericht dat bij de assertion message bijgevoegd wordt.

fail

"stop" geeft een error, "warning" geeft een warning als de assertions niet geslaagd zijn.

See also