Assert posixct named

Een functie om checkmate::assert_posixct te gebruiken om de assertion uit te voeren waarmee gecheckt wordt of de waarde een posixct type heeft.

Usage

assert_posixct_named(Kolom, df, prefix_kolom = NULL, ...)

Arguments

Kolom

een Character vector of string met de kolomnaam die getest moet worden

df

het dataframe dat de kolom bevat

prefix_kolom

standaard NULL. Als deze opgegeven wordt, wordt in de assertion message een tekst voor de variabelenaaam geplakt.

...

De overige parameters worden meegegeven aan de functie assert_posixct

See also