Bepaal doorstroom van bachelor naar master

Met bepaal_doorstroom wordt een self-join op een deel van de analyseset uitgevoerd, zodat de bachelor of premaster opleidingen en de masteropleidingen naast elkaar in een dataframe worden geplaatst. Dit dataframe bevat alleen doorgestroomde studenten.

bepaal_doorstroom(df, Opleidingsfase_vanaf, Opleidingsfase_naar, vars)

Arguments

df

Dataset waarover een self-join uitgevoerd moet worden voor verschillende opleidingsfasen. Dit is de Analyseset of een subset van de analyseset.

Opleidingsfase_vanaf

De doorstroom kan bepaald worden vanaf de bachelor of premaster, aangegeven met "B" of "P".

Opleidingsfase_naar

De doorstroom wordt bepaald naar de master, aangegeven met "M".

vars

Variabelen die meegenomen moeten worden, deze variabelen worden meegegeven in het nieuwe dataframe. De variabelen die in ieder geval meegegeven moeten worden zijn: INS_Studentnummer, INS_Opleidingsnaam_2002, INS_Opleidingsfase_BPM, INS_Studiejaar, INS_Eerste_jaar_opleiding_en_instelling. Voor bachelor naar master moet ook INS_Doorstroom_van_bachelor_naar_master meegegeven worden. Voor premaster naar master moet ook INS_Doorstroom_van_premaster_naar_master meegegeven worden.

Value

Dataframe waarbij de bachelor en masteropleidingen in dezelfde rij naast elkaar staan.

See also