build_as

A function used to update the analysis set based on an archived version and a list containing r scripts.

Usage

build_as(
  Analysis_set,
  AS_archive,
  latest,
  files_list,
  complete_rebuild = FALSE
)

Arguments

Analysis_set

the Analysis set

AS_archive

the archived version of the analysis set

latest

latest date analysis set was built

files_list

list of files to run

complete_rebuild

logical, set to TRUE to always completely rebuild AS, regardless of changes

Value

Analysis set