Build kolomwaarden

Functie om unieke waarden van kolommen te vinden, en deze, met de bijbehorende kolomnaam, weg te schrijven naar een csv bestand.

build_kolomwaarden(df, Naming, Naam_bestand)

Arguments

df

Dataframe waar de kolomnamen van geanalyseerd worden.

Naming

Een tabel of dataframe met de veldnamen van kolommen van het desbetreffende dataframe met meer informatie.

Naam_bestand

Naam van het bestand waar de kolomwaarden voor gebouwd worden, zonder bestandssoort (dus enkel "Namen" of "Gegevens")

See also