count rows Analysis set

Function to check if the number of rows of a data frame matches the specified number of rows.

count_rows_analysis_set(
  Dataset,
  ScriptName,
  Number_lines_begin,
  type = "Analysisset",
  slack = T
)

Arguments

Dataset

The dataset to be tested

ScriptName

Name The script name in which this function is executed. This one filename will be placed in message if number of rows is not equal

Number_lines_begin

The number of rows to calculate any difference

type

The name of the data file

slack

If TRUE(default), the message is also sent to Slack.