documentation add

Functie om rijen toe te voegen aan het documentatiebestand. Vervolgens wordt deze geupdate versie opgeslagen.

Usage

documentation_add(data, overwrite = F)

Arguments

data

Een data frame met de nieuwe rijen voor in het documentatiebestand

overwrite

FALSE: de rijen worden toegevoegd, zonder dat de dubbele rijen eerst verwijderd worden. Als TRUE overschrijven de nieuwe rijen de oorspronkelijke rijen.