documentation search

Doorzoekt de documentatie aan de hand van de opgegeven zoekterm. Als resultaat wordt de documentatie van de gevonden kolommen teruggegeven.

Usage

documentation_search(
  search,
  output = "veldnaam",
  omschrijving = T,
  veldnaam = T,
  categorie = T
)

Arguments

search

Zoekterm waarop de documentatie doorzocht wordt

output

"veldnaam", voor een vector met alle gevonden veldnamen. "tabel" voor een tabel met alleen Categorie, Veldnaam, en Omschrijving. "all" voor alle informatie

omschrijving

T/F indien hierin wel of niet in gezocht moet worden

veldnaam

T/F indien hierin wel of niet in gezocht moet worden

categorie

T/F indien hierin wel of niet in gezocht moet worden