Join Tkoppel isat

Een functie om opleidingsnaam (en eventueel de faculteit) uit Tkoppel_isat toe te voegen aan het dataframe. Om deze functie uit te voeren is het nodig dat in de outputfolder het volgende bestand staat: "1. Ingelezen data/INS_Tkoppel_isat.rds"

Usage

join_tkoppel_isat(df, faculteit = F)

Arguments

df

Een data frame dat de numerieke column INS_Opleidingscode_actueel bevat, en waaraan de opleidingsnaam 2002 moet worden toegevoegd.

faculteit

Als FALSE (default) wordt de column INS_Faculteit niet toegevoegd. Als TRUE wordt deze column wel toegevoegd.

Value

df