mapping Category

Analogous to 'mapping_translate', takes a csv from the documentation folder 'mapping tables'. The input of this function is a continuous value, which is mapped is moved to a category based on intervals.

mapping_category(
  Data,
  current,
  new,
  mapping_table_input = NULL,
  mapping_table_name = NULL
)

Arguments

Data

The data frame.

current

Current name.

new

New name.

mapping_table_input

Defaults to NULL. If a mapping table must be read.

mapping_table_name

Default NULL. If there is an existing mapping table must be read from the name in the folder

Value

A new data frame with new values based on the 'mapping tables'.

See also