relatief_belangrijke_waardes_columns

Een ondersteunende functie voor package_belangrijk_in_project, maar kan voor elke dataframe gebruikt worden. Met gebruik van deze functie worden columns toegevoegd die aangeven hoe belangrijk een (tel) waarde is binnen een selectie columns. De functie werkt als volgt:

Specificeer de columns waar het om gaat als 'columns' —De functie maakt de column 'sum' aan met de opsomming van alle columns —Voor elke column uit 'columns' wordt een column aangemaakt met de originele waarde uit de column gedeeld door de opsomming (column/sum) => Dit getal geeft aan hoe uniek een waarde uit een row is: 1.0 = 100% van de keren dat deze waarde voor komt is binnen deze column 0.0 = 0% van de keren dat deze waarde voor komt is binnen deze column

Usage

ratio_df_columns(
  df,
  columns = c("00. Downloaden", "01. Inlezen", "02. Manipuleren", "03. Analyseset maken",
    "04. Analyseren", "05. Rapporten")
)

Arguments

df

Dataframe

columns

De columns uit het dataframe waarover de relatieve belangrijkheid wordt berekent (en waar tevens voor elke column een extra column wordt toegevoegd)

Value

Dataframe met toegevoegde columns. In deze columns staat een waarde tussen 0.0 en 1.0. 1.0 = 100% van de keren dat deze waarde voor komt is binnen deze column 0.0 = 0% van de keren dat deze waarde voor komt is binnen deze column

Examples

if (FALSE) { # \dontrun{
dfTest <- ratio_df_columns(dfTest, c("column1", "column2"))
} # }