Read RDS CSV

a generic function that can read an R object into an RDS, fst and CSV. First determine the extension and then apply the appropriate read function

Usage

readrds_csv(
  dataloc = "",
  output,
  readr = FALSE,
  print = FALSE,
  fix.encoding = FALSE,
  encoding = "latin1",
  ...
)

Arguments

dataloc

Data location on the network drive, starting from "G:/DSZ/SA2016/Datasets/"

output

Data location on the network drive, depending on the branch set.starting from "G:/DSZ/SA2016/Datasets/Output/(branch)/"

readr

Use readr to read CSV files

print

TRUE/FALSE: Print the date on which the imported file was last modified

fix.encoding

TRUE/FALSE: Apply an encoder to the columns

encoding

Which encoding to use

...

Pass arguments for readr, read.csv2, fst or readRDS

Value

Dataframe