Retourneer de assertion message van een collectie

Geef een bericht over het wel of niet slagen van de assertion test. Hiervoor moet een "assertion collection" van het package checkmate opgegeven worden. Dit bericht wordt teruggegeven in de vorm van een error of een warning.

Usage

return_assertions_message(
  Collection,
  Name_collection,
  fail = "stop",
  silent = F,
  Outputmap = NULL
)

Arguments

Collection

Een object met de classe AssertCollection

Name_collection

De naam van de collectie. Deze naam wordt in de berichten genoemd

fail

"stop": Bij het falen van de assertions wordt een error gegeven en de uitvoer van het script gestopt. Bij "warning" wordt alleen een warning gegeven

silent

Bij FALSE (default) wordt de succesmelding in de console geprint. Bij TRUE wordt deze niet getoond

Outputmap

Map, zoals 1. Ingelezen data, waarin bestand staat

Value

De melding of de assertion-test wel of niet geslaagd is

Details

Voor enkele assertions zijn alleen warnings toegestaan, aangezien door een error een script stopt met runnen. Dit is gedaan voor de assertions: percentage missende waarden, dubbelingen, subset en set_equal.