Save FST / CSV

Function to write dataframes as fst and csv files on the shared network drive. Default behaviour is to write file as fst. Optional is to write the file as csv as well.

savefst_csv(
  Object_to_save,
  Name_to_save,
  output = "1. Ingelezen data/",
  dataloc,
  fileEncoding = "",
  save_csv = FALSE,
  session_info = FALSE,
  Test_documentatie = FALSE,
  Documentatie = FALSE,
  overwrite = TRUE
)

Arguments

Object_to_save

Het object dat opgeslagen moet worden

Name_to_save

De bestandsnaam dat het bestand moet krijgen (zonder extensie)

output

Datalocatie op de netwerkschijf, afhankelijk van de branch die is ingesteld.startend van "G:/DSZ/SA2016/Datasets/Output/(branch)/"

dataloc

Datalocatie op de netwerkschijf, startend van "G:/DSZ/SA2016/Datasets/"

fileEncoding

De fileencoding voor write.csv2 kan hier meegegeven worden. Als dit niet opgegeven wordt, wordt de standaard fileencoding gebruikt.

save_csv

Standaard FALSE: Er wordt geen CSV weggeschreven. Als TRUE wordt zowel een .rds als een .csv weggeschreven.

session_info

Standaard FALSE. Moet de sessie-info in dezelfde folder weggeschreve worden

Test_documentatie

boolean die aangeeft of test documentatie gemaakt moet worden

Documentatie

boolean die aangeeft of documentatie gemaakt moet worden

overwrite

whether to overwrite existing file when copying to other directory.