Save RDS / CSV / FST

Een generieke functie die een R object in zowel een RDS, CSV als fst opslaat.

Usage

saverds_csv(
  Object_to_save,
  Name_to_save,
  output = "1. Ingelezen data/",
  dataloc,
  fileEncoding = "",
  save_csv = FALSE,
  save_fst = FALSE,
  session_info = FALSE,
  overwrite = TRUE
)

Arguments

Object_to_save

Het object dat opgeslagen moet worden

Name_to_save

De bestandsnaam dat het bestand moet krijgen (zonder extensie)

output

Datalocatie op de netwerkschijf, afhankelijk van de branch die is ingesteld.startend van "G:/DSZ/SA2016/Datasets/Output/(branch)/"

dataloc

Datalocatie op de netwerkschijf, startend van "G:/DSZ/SA2016/Datasets/"

fileEncoding

De fileencoding voor write.csv2 kan hier meegegeven worden. Als dit niet opgegeven wordt, wordt de standaard fileencoding gebruikt.

save_csv

Standaard FALSE: Er wordt geen CSV weggeschreven. Als TRUE wordt zowel een .rds als een .csv weggeschreven.

save_fst

Standaard FALSE: Er wordt geen fst weggeschreven. Als TRUE wordt zowel een .rds als een .fst weggeschreven.

session_info

Standaard FALSE. Moet de sessie-info in dezelfde folder weggeschreve worden

overwrite

whether to overwrite existing file when copying to other directory.