Sculpt Analysis set

Wrapper rondom source() om de analyseset op te bouwen. Hierin wordt de analyseset opgegeven en de naam van het script dat gesourced moet worden. De functie maakt een kleine analyseset op basis van de volgende koppelvelden:

INS_Studentnummer

INS_Opleidingsnaam_2002

INS_Inschrijvingsjaar

INS_Studiejaar

INS_Inschrijvingsjaar_EOI

INS_Opleidingsfase_BPM

INS_ID_Student_Opleiding_Jaar

INS_ID_Student_Opleiding_Jaar_EOI

Optioneel: De velden opgegeven in Extra_sleutels

Daarnaast kunnen extra variabelen opgegeven worden als die gebruikt worden in het script met behulp van de Extra_sleutels parameter.

Usage

sculpt_analysis_set(Analysis_set, file, Extra_sleutels = NULL, slack = T)

Arguments

Analysis_set

De Analyseset

file

De locatie van het script

Extra_sleutels

Variabelenaam als textwaarde, of vector met meerdere textwaaren van variabelenamen. Dn die gebruikt worden bij het uitvoeren van het toevoegscript

slack

Bij TRUE(standaard), worden debugging berichten ook naar Slack gestuurd.