SFTP list directory

Een functie om de bestandsnamen van een map op een sftp server te verkrijgen. Alle bestanden in een folder worden opgehaald en in een vector geretourneerd. Dit is een helper functie die wordt gebruikt binnen sftp_vu_list_dir.

Usage

sftp_list_dir(remote_path, server, username, .opts)

Arguments

remote_path

Het pad van de folder op de server (exclusief server url)

server

het adres van de server, zonder protocol ervoor, en zonder slash op het einde. (dus exclusief "sftp://").

username

De gebruikersnaam die gebruikt wordt voor authenticatie op de server.

.opts

Een named lijst met opties die meegegeven worden aan de getBinaryURL functie. De opties die voor ssh authenticatie meegegeven moeten worden zijn "ssh.public.keyfile" en "ssh.private.keyfile".

Value

Een vector met de paden van alle bestanden in de opgegeven folder.