SFTP VU Upload

Functie om bestanden te uploaden naar een de VU sftp server. Het serveradres en de gebruikersnaam van StudentAnalytics zijn hier al ingevoerd. De functie gaat er van uit dat de public en private ssh key op deze locaties staan:

Functies & Libraries/ssh/User keys/sa_rsa.pub

Functies & Libraries/ssh/User keys/sa_rsa

sftp_vu_upload(local_path, remote_path)

Arguments

local_path

Het lokale pad van het bestand dat geupload wordt

remote_path

Het pad waarnaar het bestand op de server geupload moet worden (exclusief server url "Studentanalytics.sftp.vu.nl")