subassertfunctie

Functie die de meegegeven assertion checkt op basis van het meegegeven namingbestand. Dit zijn o.a. numerieke, character en date assertions.

Usage

subassertfunctie(
  naming,
  assert_functie,
  df,
  collectie,
  selecteer_kolommen,
  veldtype,
  kolomwaarden = NULL
)

Arguments

naming

documentatie met de te testen assertions

assert_functie

meegegeven assertfunctie

df

dataframe met de data die geassert wordt

collectie

assertion-collectie

selecteer_kolommen

De kolommen uit de naming die nodig zijn voor de assertion

veldtype

Veldtype van de assertion in een character of een lijst met characters

kolomwaarden

kolomwaarden om de assertions list, subset en set_equal mee te controleren