Translate colnames documentation

Function to translate the column names of a dataset to the columns contained in the documentation file are defined. The empty column names will be removed from the data.

Usage

translate_colnames_documentation(
  data,
  fieldname,
  fieldname_export,
  drop_na = T
)

Arguments

data

A dataframe whose column names need to be translated

fieldname

A column with fieldnames the dataframe will be changed to

fieldname_export

a column with the original names of the dataframe, that wil be changed

drop_na

Defaults to TRUE so that empty column names are dropped.

Value

a dataframe with translated column names