Valideer script

Een functie voor het valideren van een script met behulp van een aantal tests.

Usage

validate_script(filepath = NULL, export_naar_dataframe = FALSE)

Arguments

filepath

De map en bestandsnaam van het script dat getest moet worden binnen het project. Bijvoorbeeld: "01. Inlezen/Inlezen Afhakers.R". Als het filepath niet is opgegeven, wordt het script dat in rstudio open staat gecontroleerd

export_naar_dataframe

Logical value die aangeeft of de resultaten van de functie in een dataframe geëxporteerd moeten worden

Value

Per test wordt het resultaat weergeven. Dit gebeurd in de volgorde zoals hierboven beschreven. "Geslaagd" geeft aan dat een test zonder problemen is afgerond. Mocht dit niet het geval zijn dan geeft een waarschuwing in het rood meer informatie. De laatste output weergeeft eventuele scripts die afhankelijk zijn van het script in blauw.

Details

Het script wordt gevalideerd op het volgende:

Geeft het script waarschuwingen of errors?

Zijn indien nodig assertions aanwezig?

Worden alle variabelen op het eind van het script opgeruimd?

Voldoen alle weggeschreven bestanden aan de documentatie?

Voldoet de introductie aan de stijlgids?

Volgt het script de stijlgids?

Welke scripts zijn afhankelijk van de output bestanden van het script?

Deze functie is primair geschikt voor de inlees en manipuleer scripts. Scripts die afhankelijk zijn van eerder uitgevoerde scripts kunnen gedeeltelijk niet werken.

Let op: om de scripts te testen worden deze eenmaal uitgevoerd en worden bestanden weg geschreven. Bij sommige scripts kan deze validatie daarom wat langer duren.

See also