Write File to Project

Writes an object to a file in the project directory with specified saving options (RDS, CSV, or FST). Utilizes both manual and configuration settings for writing the file.

write_file_proj(
  object,
  name,
  settings_df = load_and_save_config_proj(),
  settings_type = NULL,
  dir = NULL,
  save_rds = NULL,
  save_csv = NULL,
  save_fst = NULL,
  sub_dir = NA_character_,
  rds_version = 2,
  csv_na = "",
  csv_sep = ";",
  csv_dec = ",",
  fst_compress = 100,
  ...
)

Arguments

object

The object to be written.

name

Character, name of the file.

settings_df

Data frame containing settings, defaults to the result of load_and_save_config_proj().

settings_type

Character, the type of setting to use, default is NULL.

dir

Character, directory path, default is NULL.

save_rds

Logical, whether to save as RDS, default is NULL.

save_csv

Logical, whether to save as CSV, default is NULL.

save_fst

Logical, whether to save as FST, default is NULL.

sub_dir

Character, sub-directory, default is NA_character_.

rds_version

Numeric, version for RDS files, default is 2.

csv_na

Character, NA representation for CSV files, default is "".

csv_sep

Character, separator for CSV files, default is ";".

csv_dec

Character, decimal separator for CSV files, default is ",".

fst_compress

Numeric, compression level for FST files, default is 100.

...

Additional arguments to be passed to the saving functions.

Value

NULL, the file is saved to the specified path.