Export analysisset.R

This function makes a subset of the analysisset. NB! A analysisset and documentation like the ones of VU student analytics is required. Because this functions uses the same columns and fieldnames.

Usage

export_analysisset(
  Faculty_selection = "All",
  Study_selection = "All",
  Phase_selection = "All",
  Academic_year = "All",
  Category_selection = NULL,
  Enrollment_year = "All",
  Enrollment_year_EOI = "All",
  Column_selection = NULL,
  Name_project,
  Special_personal_data = FALSE,
  Hash = TRUE,
  Export_results = FALSE,
  Save_as_CSV = FALSE,
  Analysis_set_input = NULL,
  Analysis_set_option = NULL,
  Documentation_new_variables = NULL,
  Qualityrapport = TRUE,
  Make_zip = FALSE,
  Save_hash = FALSE,
  Network_directory = NULL,
  Extract_location = "/6. Extracten WP/",
  Export_name = "Export_Student_Analytics_"
)

Arguments

Faculty_selection

Option between list of faculties or "All" use with format: c('faculty1','faculty2', etc)

Study_selection

Option between list of studies or "All" use with format: c('study1','study2', etc)

Phase_selection

Option between list of phases or "All". use with format: c("B", "P", "M") NB! Premaster students often have the same study name as Bachelor students, but they don't need to be selected. Keep this in mind when selecting study names.

Academic_year

Option between list of academic years or "All". use with format c(1:3), 1 or c(1,3,6).

Category_selection

Option of a list of categories. use format: c('category1','category2', etc) The standard used variables are: INS_Studentnummer, Opleiding, inschrijvingsjaar (the ID of the VUSA analysisset). Layout of the different variables in categories can be fined in the documentation of the VUSA analysisset All usable categories are: c('BSA', 'CBS', 'Functiebeperking', 'Inschrijvingen', 'Introductieweek', 'Matching', 'Middelbareschool', 'NSE', 'Orientatie', 'Resultaten', 'Succesvariabelen', 'Taaltoets', 'Voorspelmodel')

Enrollment_year

List with enrollment years, use with format: c(2013,2014)

Enrollment_year_EOI

List with enrollment years EOI, use with format: c(2013,2014)

Column_selection

If there is asked for other specific columns, then is also possible the use those. Default is NULL and their will also be looked ad category selection use with format: c('Column1','Columns2') Correct column names can be found in the documentation of the VUSA AS

Name_project

name of the exported project. use with format: "Name_project"

Special_personal_data

Option of FALSE(default) en 'Yes, etnicity' and 'yes, disability' (format: c('yes, etnicity',Ja, disability'), for both). In defalut don't deliver sensitive data. To do so: explicit permission is required

Hash

Option between TRUE (default) and FALSE.

Export_results

Option between FALSE (default) en TRUE. In addition of the analysisset, the results will also be exported.

Save_as_CSV

Option between FALSE (default) en TRUE. If TRUE not only an rds, but also a csv of the data will be saved.

Analysis_set_input

If an analysisset is made in advance, this can used. if NULL system variable for AS will be used

Analysis_set_option

Options 1, 1R, 7, or 7X

Documentation_new_variables

Variables that are not in the documentation.

Qualityrapport

Option between FALSE (default) and TRUE. If TRUE a qualityrapport will be generated

Make_zip

Option between FALSE (default) en TRUE. If TRUE a Zip of all files will be saved

Save_hash

Option between FALSE (default) en TRUE. If TRUE the Hash mapping will be saved in a map higher than the export files.

Network_directory

Location of Network

Extract_location

name of the map were extracts are saved. Default is "6. Extracten wp"

Export_name

Start of the name of your exported results Default is "Export_Student_Analytics_"

+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ TODO:

Functie herschrijven: vereenvoudigen; kolommen kunnen toevoegen

Assertion per jaar en per opleiding?

generiek maken van kolomen

+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

Details

Dependencies: Index.R Export results.R

Datasets: Output/3. Analyseset/Analysis_set_1.rds Output/2. Geprepareerde data/RES_Resultaten.rds